Move the "Ticket scanning" nav entry to the event administration
