+++ app/src/main/res/menu/event_showcase_drawer_view.xml
@@ -19,10 +19,6 @@
             android:id="@+id/nav_admin"
             android:title="@string/admin_nav_item_activity_event_showcase"
             android:visible="false"/>
-        <item
-            android:id="@+id/nav_scan"
-            android:title="@string/scan_nav_item_activity_event_showcase"
-            android:visible="false"/>
         <item android:id="@+id/nav_feedback"
             android:title="@string/feedback_nav"/>
     </group>
